update chokidar to 4.0.1
on v4 chokidar reduced dependencies to 1.
this reduces the dependency tree of squint-cljs quite a lot.

I've been using that version in a project with an override for a week
now and everything seems to work.

from the [changelog](https://github.com/paulmillr/chokidar?tab=readme-ov-file#upgrading) looks like there is nothing to be changed to update