review feedback: paths has a default of: ., src
This was only expressed internally, but needs to be also expressed in
the cli so that:
- the user can learn what the default is
- the cli uses paths itself for compile, so those defaults need to be
applied earlier for it.

This means squint no longer has any required opts.
Supporting code turfed.

This also means we no longer have our failure check for both `<files>`
and `--paths` being specified for `compile`. Updated usage note to
explain how `<files>` overrides `--paths` and `:paths`.

This allowed me to turf post-validate supporting code.